Use asyncio gather to execute all requests
